code/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdlc-11-app"
version = "0.1.0"
description = "Unknown Architecture - PDLC-11"
requires-python = ">=3.9"

dependencies = [
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.21.0",
    "python-dotenv>=1.0.0",
    "pydantic>=1.10.0",
]
